Add framework to validate Records
