--- config.json
@@ -1,4 +1,0 @@
-{
-  "lint": false,
-  "custom": "./validate.js"
-}
