+++ package.json
@@ -10,8 +10,5 @@
     "fs-extra": "^11.1.1",
     "image-size": "^1.0.2",
     "stac-node-validator": "2.0.0-beta.12"
-  },
-  "scripts": {
-    "test": "stac-node-validator --config config.json"
   }
 }
