+++ validate-records.js
@@ -1,0 +1,3 @@
+async function validateRecords(data, config, result) {
+  // TODO
+}
